test: separate ci visibility integration tests
